servers/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
sqlalchemy==1.4.41
pymysql==1.0.2
python-dotenv==0.21.0